backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Limiter==4.0.0
python-dotenv==1.0.0
pymongo==4.6.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
openai==1.3.7
boto3==1.34.0
apscheduler==3.10.4
pytest==7.4.3
pytest-flask==1.3.0
gunicorn==21.2.0
qdrant-client==1.15.1
sentence-transformers==2.2.2
torch==2.3.0
prophet==1.1.5
diskcache==5.6.3
loguru==0.7.2
plotly==5.22.0
